Created a separate IsType match operator. Restructured tests.
